split concerns off of single service
